added unit tests for cli
