+++ tests/test_cli.py
@@ -2,30 +2,22 @@
 
 """Tests for the CLI interface."""
 
-# from click.testing import CliRunner
-# from deep_translator.main import cli
-#
-#
-# class TestClass:
-#     def test_translate(self):
-#         runner = CliRunner()
-#         result = runner.invoke(cli, ['translate', 'google', '-src=auto', '-tgt=en', '-txt=좋은'])
-#         assert 'good' in result.output
-#         assert result.exit_code == 0
-#
-#     def test_api_key_error(self):
-#         runner = CliRunner()
-#         result = runner.invoke(cli, ['translate', 'microsoft','-src=auto','-tgt=en','-txt=\'Zwei minimale Dellchen auf der Rückseite.\''])
-#         assert "This translator requires an api key provided through --api-key" in result.output
-#         assert result.exit_code == 1
-#
-#     def test_language_languages(self):
-#         runner = CliRunner()
-#         result = runner.invoke(cli, ['languages', 'google'])
-#         assert result.exit_code == 0
-#
-#     def test_invalid_language_languages(self):
-#         runner = CliRunner()
-#         result = runner.invoke(cli, ['languages', 'notValidTranslator'])
-#         assert 'The given translator is not supported.' in str(result.exception)
-#         assert result.exit_code == 1
+from deep_translator.cli import CLI
+import pytest
+import sys
+
+
+@pytest.fixture
+def mock_args():
+    sys.argv[1:] = ['--source', 'en', '--target', 'de', '--text', 'hello']
+    return CLI(sys.argv[1:]).parse_args()
+
+
+def test_source(mock_args):
+    assert mock_args.source == 'en'
+
+
+def test_target(mock_args):
+    assert mock_args.target == 'de'
+
+
